+++ src/Container.php
@@ -119,6 +119,11 @@ class Container implements ContainerInterface
 
     /**
      * @inheritDoc
+     * @template T
+     *
+     * @param class-string<T> $id
+     *
+     * @return T
      */
     public function get(string $id): mixed
     {