+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Compatibility with `psr/container` version 2
 - "Nullable" services
+- "@template" annotation for main container
 
 ## [2.0.3] - 2021-09-23